requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Ethereum/Monad integration
web3==6.15.0
eth-account==0.10.0
py-solc-x==2.0.2

# HTTP client
httpx==0.26.0

# AI/LLM integration
anthropic==0.18.1

# Utilities
python-dotenv==1.0.0
keyring==25.6.0
